legacy(py)/requirements.txt
fastapi
uvicorn
python-multipart
requests
pyyaml
openai
pydantic
litellm
pytest
yfinance
mcp
bcrypt
pandas
numpy
Wikipedia-API
youtube-transcript-api
mutagen
spotipy
selenium
beautifulsoup4
httpx
mem0ai
google-genai
Pillow
python-dotenv
email-validator
pyjwt
